refactor 'components' directory separated pages and components into respective directories
